+++ Mu3Library_URP/Samples~/ScreenEffect/Scripts/Runtime/VolumeHandle/ShakeHandler.cs
@@ -1,4 +1,5 @@
 using Mu3Library.URP.ScreenEffect.Effects.Shake;
+using UnityEngine;
 
 namespace Mu3Library.URP.Sample.ScreenEffect.VolumeHandle
 {
@@ -11,19 +12,25 @@ namespace Mu3Library.URP.Sample.ScreenEffect.VolumeHandle
         {
             base.OnInit();
 
-            _effect.SetWeight(0f);
-            _effect.SetAmplitude(0f);
+            SetValueWeight(0f);
+            SetValueAmplitude(0f);
+            SetValuePeriod(0f);
         }
 
         #region UI Event
         public void SetValueWeight(float value)
         {
-            _effect?.SetWeight(value);
+            _effect?.SetWeight(Mathf.Lerp(ShakePass.WeightMin, ShakePass.WeightMax, value));
         }
 
         public void SetValueAmplitude(float value)
         {
-            _effect?.SetAmplitude(value);
+            _effect?.SetAmplitude(Mathf.Lerp(ShakePass.AmplitudeMin, ShakePass.AmplitudeMax, value));
+        }
+
+        public void SetValuePeriod(float value)
+        {
+            _effect?.SetPeriod(Mathf.Lerp(ShakePass.PeriodMin, ShakePass.PeriodMax, value));
         }
         #endregion
     }